fix(pr): address comments
